add monitor server config in shared app struct
