+++ rust/numaflow-monitor/src/lib.rs
@@ -8,8 +8,16 @@ use crate::config::generate_certs;
 use crate::config::server::MonitorServerConfig;
 use crate::error::{Error, Result};
 use axum_server::tls_rustls::RustlsConfig;
+use config::info::RuntimeInfoConfig;
+use runtime::Runtime;
 use std::net::SocketAddr;
+use std::sync::{Arc, RwLock};
 use tracing::info;
+
+pub(crate) struct AppState {
+    pub(crate) runtime: Arc<RwLock<Runtime>>,
+    pub(crate) server_config: MonitorServerConfig,
+}
 
 pub async fn run() -> std::result::Result<(), Box<dyn std::error::Error + Send + Sync + 'static>> {
     let (cert, key) = generate_certs()?;
@@ -19,6 +27,14 @@ pub async fn run() -> std::result::Result<(), Box<dyn std::error::Error + Send +
         .map_err(|e| format!("Failed to create tls config {:?}", e))?;
 
     let server_config = MonitorServerConfig::default();
+    let runtime_info_config = RuntimeInfoConfig::default();
+    let runtime = Runtime::new(runtime_info_config.app_error_path.as_str());
+
+    // Initialize shared state
+    let shared_state = Arc::new(AppState {
+        runtime: Arc::new(RwLock::new(runtime)),
+        server_config: server_config.clone(),
+    });
 
     info!(?server_config, "Starting monitor server with config");
 
@@ -27,7 +43,7 @@ pub async fn run() -> std::result::Result<(), Box<dyn std::error::Error + Send +
         .parse()
         .map_err(|e| Error::Init(format!("{e:?}")))?;
     // Start the main server, which serves the application.
-    start_main_server(app_addr, tls_config, server_config).await?;
+    start_main_server(app_addr, tls_config, shared_state).await?;
 
     Ok(())
 }
